refactor: move parse function to its own file
